+++ docs/changelog.rst
@@ -13,9 +13,9 @@ news
 
 - Now the standard deviation in thresholds by below (for 3 or 7 categories) need
   specify the negative values, e.g: sd-2.1, sd-1.2, sd-0.5, sd0.6, sd0.9, sd1.3
-- Defined internal calculation_mode_series for var D and I in 'default' value
+- Defined internal mode_calculation_series for var D and I in 'default' value
   and check it when is chose by the user
-- New options in runfile: calculation_mode_series_D and calculation_mode_series_I,
+- New options in runfile: mode_calculation_series_D and mode_calculation_series_I,
   these are the options to calculate the series: accumulate or mean
 - Updated thresholds for var D for 3 and 7 categories (ATTENTION: this change
   modify several results)